requirements.txt
beautifulsoup4==4.9.3
Django==2.2.16
django-bootstrap-datepicker==1.4.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-cleanup==5.1.0
et-xmlfile==1.0.1
importlib-metadata==1.7.0
jdcal==1.4.1
openpyxl==3.0.5
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.16.0
soupsieve==2.2
sqlparse==0.3.1
xlrd==2.0.1
zipp==3.2.0
